fix: address phase2 review findings
